fix: moved vidazoo buildRequests to lib
